Major reshuffling of files and directories
