+++ src/paradigma/pipelines/gait_pipeline.py
@@ -106,36 +106,6 @@ def extract_gait_features(
     return df_features
 
 
-def extract_gait_features_io(
-        config: GaitConfig,
-        path_to_input: str | Path, 
-        path_to_output: str | Path
-    ) -> None:
-    # Load data
-    metadata_time, metadata_values = read_metadata(path_to_input, config.meta_filename, config.time_filename, config.values_filename)
-    df = tsdf.load_dataframe_from_binaries([metadata_time, metadata_values], tsdf.constants.ConcatenationType.columns)
-
-    # Extract gait features
-    df_features = extract_gait_features(df=df, config=config)
-
-    # Store data
-    end_iso8601 = get_end_iso8601(start_iso8601=metadata_time.start_iso8601,
-                                  window_length_seconds=int(df_features[DataColumns.TIME][-1:].values[0] + config.window_length_s))
-
-    metadata_values.file_name = 'gait_values.bin'
-    metadata_time.file_name = 'gait_time.bin'
-    metadata_values.end_iso8601 = end_iso8601
-    metadata_time.end_iso8601 = end_iso8601
-    
-    metadata_values.channels = list(config.d_channels_values.keys())
-    metadata_values.units = list(config.d_channels_values.values())
-
-    metadata_time.channels = [DataColumns.TIME]
-    metadata_time.units = [TimeUnit.RELATIVE_S]
-
-    write_df_data(metadata_time, metadata_values, path_to_output, 'gait_meta.json', df_features)
-
-
 def detect_gait(
         df: pd.DataFrame, 
         clf_package: ClassifierPackage, 
@@ -187,39 +157,6 @@ def detect_gait(
     pred_gait_proba_series = clf_package.predict_proba(X)
 
     return pred_gait_proba_series
-
-
-def detect_gait_io(
-        config: GaitConfig, 
-        path_to_input: str | Path, 
-        path_to_output: str | Path, 
-        full_path_to_classifier_package: str | Path, 
-    ) -> None:
-    
-    # Load the data
-    config.set_filenames('gait')
-
-    metadata_time, metadata_values = read_metadata(path_to_input, config.meta_filename, config.time_filename, config.values_filename)
-    df = tsdf.load_dataframe_from_binaries([metadata_time, metadata_values], tsdf.constants.ConcatenationType.columns)
-
-    clf_package = ClassifierPackage.load(full_path_to_classifier_package)
-
-    df[DataColumns.PRED_GAIT_PROBA] = detect_gait(
-        df=df, 
-        clf_package=clf_package
-    )
-
-    # Prepare the metadata
-    metadata_values.file_name = 'gait_values.bin'
-    metadata_time.file_name = 'gait_time.bin'
-
-    metadata_values.channels = [DataColumns.PRED_GAIT_PROBA]
-    metadata_values.units = ['probability']
-
-    metadata_time.channels = [DataColumns.TIME]
-    metadata_time.units = [TimeUnit.RELATIVE_S]
-
-    write_df_data(metadata_time, metadata_values, path_to_output, 'gait_meta.json', df)
 
 
 def extract_arm_activity_features(
@@ -366,66 +303,6 @@ def extract_arm_activity_features(
     return df_features
 
 
-def extract_arm_activity_features_io(
-        config: GaitConfig, 
-        path_to_timestamp_input: str | Path, 
-        path_to_prediction_input: str | Path, 
-        full_path_to_classifier_package: str | Path, 
-        path_to_output: str | Path
-    ) -> None:
-    # Load accelerometer and gyroscope data
-    dfs = []
-    for sensor in ['accelerometer', 'gyroscope']:
-        config.set_sensor(sensor)
-        meta_ts_filename = f'{sensor}_meta.json'
-        values_ts_filename = f'{sensor}_values.bin'
-        time_ts_filename = f'{sensor}_time.bin'
-
-        metadata_ts_dict = tsdf.load_metadata_from_path(os.path.join(path_to_timestamp_input, meta_ts_filename))
-        metadata_ts_time = metadata_ts_dict[time_ts_filename]
-        metadata_ts_values = metadata_ts_dict[values_ts_filename]
-        dfs.append(tsdf.load_dataframe_from_binaries([metadata_ts_time, metadata_ts_values], tsdf.constants.ConcatenationType.columns))
-
-    df_ts = pd.merge(dfs[0], dfs[1], on=DataColumns.TIME)
-
-    # Load gait predictions
-    meta_pred_filename = 'gait_meta.json'
-    values_pred_filename = 'gait_values.bin'
-    time_pred_filename = 'gait_time.bin'
-
-    metadata_pred_dict = tsdf.load_metadata_from_path(os.path.join(path_to_prediction_input, meta_pred_filename))
-    metadata_pred_time = metadata_pred_dict[time_pred_filename]
-    metadata_pred_values = metadata_pred_dict[values_pred_filename]
-
-    df_pred_gait = tsdf.load_dataframe_from_binaries([metadata_pred_time, metadata_pred_values], tsdf.constants.ConcatenationType.columns)
-
-    clf_package = ClassifierPackage.load(full_path_to_classifier_package)
-
-    # Extract arm activity features
-    config = GaitConfig(step='arm_activity')
-    df_features = extract_arm_activity_features(
-        config=config,
-        df_timestamps=df_ts, 
-        df_predictions=df_pred_gait, 
-        threshold=clf_package.threshold
-    )
-
-    end_iso8601 = get_end_iso8601(metadata_ts_values.start_iso8601, df_features[DataColumns.TIME][-1:].values[0] + config.window_length_s)
-
-    metadata_ts_values.end_iso8601 = end_iso8601
-    metadata_ts_values.file_name = 'arm_activity_values.bin'
-    metadata_ts_time.end_iso8601 = end_iso8601
-    metadata_ts_time.file_name = 'arm_activity_time.bin'
-
-    metadata_ts_values.channels = list(config.d_channels_values.keys())
-    metadata_ts_values.units = list(config.d_channels_values.values())
-
-    metadata_ts_time.channels = [DataColumns.TIME]
-    metadata_ts_time.units = [TimeUnit.RELATIVE_S]
-
-    write_df_data(metadata_ts_time, metadata_ts_values, path_to_output, 'arm_activity_meta.json', df_features)
-
-
 def filter_gait(
         df: pd.DataFrame, 
         clf_package: ClassifierPackage, 
@@ -470,38 +347,6 @@ def filter_gait(
     pred_no_other_arm_activity_proba_series = clf_package.predict_proba(X)
 
     return pred_no_other_arm_activity_proba_series
-
-
-def filter_gait_io(
-        config: GaitConfig, 
-        path_to_input: str | Path, 
-        path_to_output: str | Path, 
-        full_path_to_classifier_package: str | Path, 
-    ) -> None:
-    # Load the data
-    config.set_filenames('arm_activity')
-
-    metadata_time, metadata_values = read_metadata(path_to_input, config.meta_filename, config.time_filename, config.values_filename)
-    df = tsdf.load_dataframe_from_binaries([metadata_time, metadata_values], tsdf.constants.ConcatenationType.columns)
-
-    clf_package = ClassifierPackage.load(filepath=full_path_to_classifier_package)
-
-    df[DataColumns.PRED_NO_OTHER_ARM_ACTIVITY_PROBA] = filter_gait(
-        df=df, 
-        clf_package=clf_package
-    )
-
-    # Prepare the metadata
-    metadata_values.file_name = 'arm_activity_values.bin'
-    metadata_time.file_name = 'arm_activity_time.bin'
-
-    metadata_values.channels = [DataColumns.PRED_NO_OTHER_ARM_ACTIVITY_PROBA]
-    metadata_values.units = ['probability']
-
-    metadata_time.channels = [DataColumns.TIME]
-    metadata_time.units = [TimeUnit.RELATIVE_S]
-
-    write_df_data(metadata_time, metadata_values, path_to_output, 'arm_activity_meta.json', df)
 
 
 def quantify_arm_swing(
